WIP: Architecture refactor
