+++ app/src/main/java/com/ivy/wallet/ui/theme/components/IvyColorPicker.kt
@@ -22,12 +22,12 @@ import androidx.compose.ui.unit.dp
 import com.ivy.design.api.navigation
 import com.ivy.design.l0_system.UI
 import com.ivy.design.l0_system.style
+import com.ivy.fp.test.TestingContext
 import com.ivy.wallet.R
 import com.ivy.wallet.ui.IvyWalletComponentPreview
 import com.ivy.wallet.ui.ivyWalletCtx
 import com.ivy.wallet.ui.paywall.PaywallReason
 import com.ivy.wallet.ui.theme.*
-import com.ivy.wallet.utils.TestingContext
 import com.ivy.wallet.utils.densityScope
 import com.ivy.wallet.utils.onScreenStart
 import com.ivy.wallet.utils.thenIf
